Gcc/Glibc make will patch in the -ros files
When rebuilding gcc/glibc after commits that add or modify files in the
-ros directories (e.g. glibc-XXX-ros), you no longer need to worry about
copying the changed files from the -ros directories into the main
directory; the make process will patch it for you.

One thing it won't do is restore files that were overwritten by -ros
files if the -ros file is subsequently removed.  Meaning, if a commit
removes files from a -ros directory, you need to either blow away the
main directory (glibc-XXX, for instance), or extract the original
version of the specific file from the tarball and put it in the
appropriate place in the main directory.