Optimizes run_once(_safe) (XCC)
Takes the old run_once_safe and optimizes it for being completed, so we
don't hit an atomic_swap (and a couple branches) each time through.
Given this, I don't mind using it in more places, so the non-safe
version is now the 'racy' version (same with init_once()).

Reinstall your kernel headers.