Inlines spinlocks when not using SPINLOCK_DEBUG
This one's for Andrew.  =)  Additionally, this moves the responsibility
for memory barriers to the architectures, which cuts down on the number
of full mb()s needed by RISCV (esp since Andrew was already doing one in
there).

Also, I imagine sparc wasn't compiling, so this might work for them too,
if anyone still uses it.