Moves futex.h to user/pthread (XCC)
As far as I can tell, libgomp isn't even building with the ros/bar.c or
ros/futex.h (garbage in those files still compiles), and if it does for
some configuration, then libgomp.a doesn't have futex as an undefined
symbol, which is a bad configuration.

For now, just move it out of the kernel headers, since it isn't helping
much there at all.

The linux version of futex.h seems to think we need to change the
visibility too, so putting that in for now.

Reinstall your kernel headers.  I'd also remove the old ros/futex.h
manually.