Removes the -ros folder updating in gcc/glibc(XCC)
The old change doesn't work well with the riscv patches.  Until one day
when we get rid of the -ros directories and have separate repos for
glibc/gcc, we'll stick with the current style:

If there are changes to files in the -ros directory and you don't want
to make clean / whatever, then you need to manually copy those files
over from the -ros directory to the untarred directory (e.g.
gcc-4.6.1/).