Lock depth checking allows panicy prints
We use locks when printing, so once one lock failed, it would
recursively fail if the condition remained true (such as having a bad
kernel stack depth).  This would happen for any kernel trap or trace
coretf, or any other place where I don't care about the deadlock but
want to get the printf.

Note that there are a couple locks involved in printing - two on the
printk path.  If we fail (e.g. for IRQSAVE reasons) on the SECOND lock
(after grabbing the first lock), we will still deadlock, since we
basically grabbed lock 1, failed when getting 2, then attempted to get 1
again for another printk.  In lieu of making this debug code more
robust, we just will leave those locks as irqsave for a long time.