Fixes bug with pop_ros_tf
Removing memset() in 0a19848d0 was a little hasty.  We don't need to
zero out the entire struct, but we did need to have zeroed flags.  o/w,
the kernel might see the SC_UEVENT flag set, and think userspace wants
an event sent to whatever gibberish is on the stack at sysc->ev_q.