Spin-PDR locks (preemption detection and recovery)
Two variants: with and without CAS.  You can control which one you get
via the CONFIG var in your makelocal.  The default is to use CAS, and
RISCV will get NO_CAS (in spinlock.h) regardless of what you ask for.

The main drawback to the NO_CAS variety is the race where a lockholder
isn't advertised yet, and we're forced to do a linear scan of the
vcoremap to make sure all possible vcores are not preempted.  It won't
deadlock, but it could suck (though it might just be a minor
inconvenience compared to being preempted).