Updates doc with new image file and instructions
I built a new, larger virtual machine image file with grub and akaros
preloaded:
http://akaros.cs.berkeley.edu/files/hdd268mb.img

Or follow the updated instructions on how to make your own.  The image
will actually be smaller (in space consumed) if you make it yourself,
and will only grow up to 268MB if you use all of that space.