Handle rdtscp
rdtscp has a bunch of benefits for measurements (stay tuned for more
commits), but it only works on more recent x86 hardware, and doesn't
work on my old core2 development machine running qemu.

x86 will now catch the rdtscp invalid opcode trap and fake the command.
This will not give accurate numbers for measurement.  During boot, look
for whether or not RDTSCP is supported if you are curious.

In the future, we could have a flag that replaces rdtscp calls with
rdtsc at compile time, if someone wanted more accurate measurements on
older hardware.  For now, I'm not too concerned with those older
machines when it comes to measuring performance.