Changes rdtsc serialization methods
In short, use lfence;rdtsc; instead of cpuid;rdtsc.  For the long
version, read the info at the beginning of k/a/i/rdtsc_test.c.

Also, this will attempt to detect whether or not your hardware has any
issues preventing the reliable use of a TSC.  This includes not having
an invariant TSC, as well as getting SMI interrupts.  Again, read
rdtsc_test.c