x86: handle overflow when setting the lapic timer
Note that if you have a long-running alarm and a short periodic alarm
that goes off frequently, the long-runner will continue to be reset
every time the periodic runs off.  For a brief moment, the long-runner
will be the next up, which will cause an overflow calculation, til the
periodic resets itself (depending on the nature of the periodic reset.
The ksched does the reset outside of the handler, for instance).