Kernel messages no longer take *tf
I've been considering getting rid of this for a while - it is mostly an
artifact of old-school KMSGs that operated on running TFs, which was
from back before we were tracking the user-tf via pcpui->cur_tf.  Those
old messages wanted to operate on the user context, not whatever
hardware TF happened to be running at the time the KMSG executes.