Changes hardware trapframe calls to use new struct
All places where we used a struct trapframe and that was a HW trapframe
(compared to a SW trapframe or user_context) now use the new struct
hw_trapframe.

I considered removing the TF from the monitor calls, since I don't see
much use in it.  Monitor functions that want to use it can't assume it
is there (only ever called from breakpoint).  Maybe we'll find a
debugging use for it.