x86: initialize XMM registers along with x87
In the future (or in RISCV), someone might have an init_fp_state() that
is super-efficient.  If it turns out that manually zeroing the XMM
registers is faster, then we can change the code.  I haven't
microbenchmarked the new way or compared it to 0'ing the 15 XMM
registers + loading the MXCSR yet.