Adds syscall for help entering VC context (XCC)
We had been using sys_self_notify for this, but I actually want
something that doesn't require a specific vcoreid and just acts on the
calling pcore.

There are a couple cases with the old style where we could send
(harmless) notifs to the wrong core.  If the uthread were interrupted
and migrated right after it decided to self_notify, when it was repopped
in the future, it would notify its old vcore.

The other nice thing about this is that we don't need to worry about the
storage space for the syscall, simplifying our pop code a bit.  I didn't
bother removing the local_sysc from the restart_helper.  Not worth the
hassle, and I'll do it on the next rewrite of the HW TF stuff.

Reinstall kernel headers, rebuild parlib + apps.