Allows XCC building with newer texinfo (XCC)
binutils and stage2 of gcc are building some documentation that has a
few bugs:
http://sourceware.org/bugzilla/show_bug.cgi?id=15183

We just disable the building, using the method described in
http://wiki.osdev.org/Talk:GCC_Cross-Compiler

If you've been purposefully using old versions of texinfo, please
upgrade and test this out.  Works on my gentoo box.  You'll need to rm
the builddirs and hidden configure files in tools/compilers/gcc-glibc/.