Forces all applications to link with -lm (XCC)
Since parlib's measure has math functions in it, anyone linking against
parlib will need -lm.  Since we're forcing -lparlib, we might as well do
-lm too.  The alternative would be to not allow math functions in
parlib.  While I can get by with that for now (move measure somewhere
else or rewrite sqrt()), I can imagine us using math in parlib in the
future.

If it becomes a big deal, we can revert this.  Though libm.so is less
than 10% of the size of libc.so (660K vs 7.5M).

Rebuild gcc fully.  Make clean, if you are unsure of what to do.