x86: linker script uses sections, not names
It's a bit cleaner, and I'll need this for the upcoming Kbuild patches.

As a side note, the boot section is three pages.  One for the code, and
one for the page table, but also one for the GDT data.  I'd like to
include that in the same page as the code, but I guess the linker
doesn't like that.  I guess we could ultimately free up those three
pages once we've fully booted, if we're hurting for 12KB.