Initial Kbuild / Kconfig commit
This is just the contents of git://github.com/lacombar/kconfig.git,
which was made back in Aug 2011.  It should be a slimmed down version of
the Kbuild essentials.

The one modification I made was to move Makefile -> Makefil-KB for now.

In case we ever update it, I'll keep my changes in a separate commit.