Reverts -lm error (XCC)
This fixes the change from 6f3a1535f7, where all applications linked
with -lm.  In lieu of libm, parlib won't use any math, and whatever
utilities I come up with will just get dumped in benchutil.

The problem was that the cross-compiler can't bootstrap itself if there
was not an existing installation with -lm.  It would work only after
everything was installed, and it took a reinstall with no existing
toolchain to detect it.

The make process for benchutil was just a copy of parlib's.  We probably
ought to take a better look at all of userspace's makes, including where
we drop files in the toolchain directories.

If you have compilation issues, make clean and rm your toolchain
installation.