Modified the build system to use Kconfig
Once we have entries in the Kconfig, the build system will pass any
CONFIG_FOO=y into the C files as #define CONFIG_FOO 1.  Whenever any
config variable changes, any file using that #define will be rebuilt.

Also, we support the silentoldconfig recursive target, which detects
changed .config files and regenerates the kconfig outputs, so that
kbuild will detect the changes.

The downside to the silent rebuild of auto.conf is that it exposed some
bug in objdump where we can't use a -S when we also updated and included
auto.conf.  That was a pain to figure out.  If you really want -S, you
can do it, so long as you don't rely on the silentoldconfig recursive
call.  You can do something like:

$ touch .config
$ make silentoldconfig
$ make (with -S in your objdump args)