Adds 64 bit option
Had to move the CROSS_COMPILE detection section to after including auto.conf,
so that we get up-to-date values for it.  Otherwise, when switching bitness, we
could get the same cross compiler as the previous run.

This also gets rid of the old $(error) when we can't find the cross compiler.
Instead, we detect it when we bulid.  The clean targets shouldn't rely on those
XCC related values anyways, but we'll see.

This also auto-includes common.h (and also ros/common.h) into every kernel
file.