x86: objdump 64 bit kernel properly
The kernel image's format is a 32 bit elf to trick grub into loading it.
This also tricks objdump into interpreting the asm as if it was 32 bit.
The -M tells objdump what the real format is.

Incidentally, now that I can see the asm, I can see the compiler does
generate xmm code in some places, and we do in fact need the extra
cflags.