Fixes page-faulting glibc on certain machines (XCC)
The i686 toolchain when built on brho's desktop was causing apps linked
against glibc to page fault in __flockfile (_IO_flockfile) on
(*stream->lock).owner.  I'm not sure why this was limited to certain
machines, or if we've seen the end of this bug.

You'll need a full rebuild of the toolchain (at least of glibc).