x86_64: all apps compiled with -mno-red-zone (XCC)
The red-zone is an ABI specification, allowing the compiler to use 128
bytes below the stack pointer.  This helps with some leaf functions.
We make a bunch of assumptions about being able to use the area below
the stack for popping user contexts, so we can't allow the red-zone
optimization.  There is a similar issue with interrupts in the kernel
(given our current interrupt handling strategy).

Copy over ros64.h and rebuild GCC.