Changes Akaros default build to x86_64
If you want something else, you'll have to do a small amount of fiddling.  To
mess with the kernel/userspace config, menuconfig will do the trick.  You can
do the same for busybox, when in their source tree, or just edit the .config.

For the cross compiler, just make the appropriate target (even for x86_64).