x86_64: support for machines with only 2MB pages
Some machines (future AMD ones? and definitely some VMs) don't support 1
GB jumbo pages.

This adds in support for using 2MB jumbos for the boot mappings.

This adds an extra 2MB or so of extra memory in the data segment for the
pml2s, regardless of whether or not your machine needs them.  I might
add a CONFIG for that at some point.