Add basic timeout functionality to pthread futexes
In the current implementation we simply launch a helper pthread that
calls sys_block(1000) in a tight loop.  It wakes up every millisecond
and checks if there are any futexes whose timeouts have expired and it
wakes them up. If there are no futexes waiting on a timer, we park the
pthread and wait until another futex is waited on that has timeout.