x86: catches buggy rdtscp
Not sure if this is a "bug in qemu" or something I'm doing wrong.  But rdtscp
should drop the value of MSR_TSC_AUX in ecx.  If this fails, it busts our
FAST_COREID.  If this happens to you in a VM, just change your CONFIG.

If it happens on hardware, let me know.