Vcore TLS desc in procdata (XCC)
Userspace now tells the kernel what TLS descriptor to use for each vcore.  The
main reason is that when making a new context, the kernel would zero the TLS
base, and then userspace would quickly set it to the vcore's TLS.  On amd64,
this incurs an extra syscall (for now), and on i686, it's an extra 70ns for the
lousy-LDT mangling.  This change allows the kernel to quickly set the TLS base
(and optionally not change it if it is already correct).

Note that only amd64 uses this option in the kernel; riscv and i686 still
handle this in userspace, but also expose their tlsdesc to the kernel.

The other benefit to this is debugging - it's often helpful for me to have
access to this from the kernel.

As a historic note, this isn't the first arch-dependent tidbit in procdata.
The transition_stack (vcore's stack) was added to procdata since sparc couldn't
handle not having a stack (x86 was just loading %sp in assembly before _start).