Userspace handling of errstr and errno (XCC)
Since not all uthreads have TLS, errno and the new errstr need to be
decoupled from the TLS.

When we're a uthread, we'll use storage in the uthread struct for errno
and errstr.  Otherwise, we'll default to the TLS locations (for vcores).
ld.so will continue to use its own area.  Very early in start.c, before
we call out to uthread code, we'd also use the TLS location.

perror() will print out the errstr, if there is one like so:

"your-message: errno-translated-to-text, errstr"

Rebuild your cross compiler.