Fixes bug in KFS new inode creation
Wasn't properly initializing the size field.  This would only matter
when creating new files in KFS, and would only trigger if the memory we
alloc'd wasn't already zeroed.

Something like      $ echo "" >> tmpfile    was enough to cause it, once
the bits lined up right.

The previous commit's backtrace was rather useful in determining this
(without it, all we know is there was a KPF in memcpy).