x86: option to disable the keyboard
To avoid some SMM overheads on machines with USB keyboards.

Without this, simply sitting in the kernel monitor on core 0 has a massive
performance hit on some benchmarks (10-15x for fluidanimate).

Note that sitting in the monitor still has an effect on performance, for now,
even with the KB disabled.