Namec works. Waserror is way improved.
It's not quite as easy to use as Plan 9 but waserror is way better.
It also now has a great deal of integrity checking at the cost of a bit more
work on the programmer's part.

And, the 9open test works :-)

Changes to date and issues:
changing incref/decref to krefs (semantics)
removing utf-8
Translating proc structs
The fact that we don't have current-> or up->

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>