Make plan 9 fd range be the same as normal FDs
I underestimated how important it was to busybox to have
fds in the range 0 ...

For example, if you do
dd of=x
it wants to dup x over 1 and then write to 1. Klever.

So, we add an additional ghetto hack in vfs so we can tell plan 9
files apart and move forward.

N.B. I am aware this is an ugly hack but
this not a situation I expect to hold for very long. If I get
this right Barret and Kevin will let us move to the plan 9 stuff as
the main show and we can kill this compatibility mode.

cat '#Z/regressctl' works but cat '#Z' fails. I'll fix that later today
I hope.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>