Fix namec
Probably due to a utf-8 removal or something, devwalk was allocating bizarre
amounts of data (nname-1) * etc. when nname was 0.

Thank you barret, I will sleep better.

This is the start of cleanup. Next step is an lindent pass, and then we
start looking at a merge.

Devvirtio, devproc, here we come

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>