Fixes up lindent
indent still messes up sometimes.  In some cases (like with weird printk
formatting), it can be fixed and won't cause the problem again.  In
others, like in sys_buster right after the #defines, it'll insist on
formatting in a weird manner.

Anyway, this is better than not formatting the code at all.