Add the pipe device
This caught some wrong bits in the various functions
related to perrbuf.

Does it work? I can't tell. But if we get this going
devether will also be easier.

Also, change the name from | to P. The busybox shell can't deal with
the pipe symbol even when it's in quotes.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>